Remove db references from API (#61)
Moves more DB logic into the Common library pulling logic out of the API
handlers.
