+++ packages/common/src/services/posts/index.ts
@@ -1,3 +1,6 @@
 export * from './listPosts';
 export * from './getPosts';
 export * from './createPost';
+export * from './deletePost';
+export * from './listRelatedOrganizationPosts';
+export * from './createPostInOrganization';
