+++ packages/common/src/services/reactions/addReaction.ts
@@ -1,0 +1,30 @@
+import { and, db, eq } from '@op/db/client';
+import { postReactions } from '@op/db/schema';
+
+export interface AddReactionOptions {
+  postId: string;
+  profileId: string;
+  reactionType: string;
+}
+
+export const addReaction = async (options: AddReactionOptions) => {
+  const { postId, profileId, reactionType } = options;
+  await db.transaction(async (tx) => {
+    // First, remove any existing reaction from this user on this post
+    await tx
+      .delete(postReactions)
+      .where(
+        and(
+          eq(postReactions.postId, postId),
+          eq(postReactions.profileId, profileId),
+        ),
+      );
+
+    // Then add the new reaction
+    await tx.insert(postReactions).values({
+      postId,
+      profileId,
+      reactionType,
+    });
+  });
+};
