+++ packages/common/src/services/reactions/index.ts
@@ -1,1 +1,5 @@
 export * from './validateReaction';
+export * from './addReaction';
+export * from './removeReaction';
+export * from './getExistingReaction';
+export * from './toggleReaction';
