+++ packages/common/src/services/reactions/validateReaction.ts
