+++ packages/ui/tailwind.shared.ts
@@ -140,38 +140,6 @@ const config: Omit<Config, 'content'> = {
       colors: {
         accent: commonColors,
         border: 'hsl(var(--op-offWhite))',
-        // input: 'hsl(var(--input))',
-        // ring: 'hsl(var(--ring))',
-        // background: 'hsl(var(--background))',
-        // foreground: 'hsl(var(--foreground))',
-        // primary: {
-        //   DEFAULT: 'hsl(var(--primary))',
-        //   foreground: 'hsl(var(--primary-foreground))',
-        // },
-        // secondary: {
-        //   DEFAULT: 'hsl(var(--secondary))',
-        //   foreground: 'hsl(var(--secondary-foreground))',
-        // },
-        // destructive: {
-        //   DEFAULT: 'hsl(var(--destructive))',
-        //   foreground: 'hsl(var(--destructive-foreground))',
-        // },
-        // muted: {
-        //   DEFAULT: 'hsl(var(--muted))',
-        //   foreground: 'hsl(var(--muted-foreground))',
-        // },
-        // accent: {
-        //   DEFAULT: 'hsl(var(--accent))',
-        //   foreground: 'hsl(var(--accent-foreground))',
-        // },
-        // popover: {
-        //   DEFAULT: 'hsl(var(--popover))',
-        //   foreground: 'hsl(var(--popover-foreground))',
-        // },
-        // card: {
-        //   DEFAULT: 'hsl(var(--card))',
-        //   foreground: 'hsl(var(--card-foreground))',
-        // },
         teal: {
           DEFAULT: 'hsl(var(--op-teal-500))',
           50: 'hsl(var(--op-teal-50))',
