+++ services/api/src/routers/account/getUserProfiles.ts
@@ -1,10 +1,9 @@
-import { UnauthorizedError } from '@op/common';
+import { UnauthorizedError, getUserWithProfiles } from '@op/common';
 import { EntityType, ObjectsInStorage, Profile } from '@op/db/schema';
 import type { OpenApiMeta } from 'trpc-to-openapi';
 import { z } from 'zod';
 
 import withAuthenticated from '../../middlewares/withAuthenticated';
-import withDB from '../../middlewares/withDB';
 import withRateLimited from '../../middlewares/withRateLimited';
 import { loggedProcedure, router } from '../../trpcFactory';
 
@@ -37,7 +36,6 @@ export const getUserProfiles = router({
   getUserProfiles: loggedProcedure
     .use(withRateLimited({ windowSize: 10, maxRequests: 10 }))
     .use(withAuthenticated)
-    .use(withDB)
     .meta(meta)
     .input(z.undefined())
     .output(
@@ -53,33 +51,10 @@ export const getUserProfiles = router({
       ),
     )
     .query(async ({ ctx }) => {
-      const { db } = ctx.database;
       const { id: authUserId } = ctx.user;
 
       // Get the user's database record
-      const user = await db.query.users.findFirst({
-        where: (table, { eq }) => eq(table.authUserId, authUserId),
-        with: {
-          profile: {
-            with: {
-              avatarImage: true,
-            },
-          },
-          organizationUsers: {
-            with: {
-              organization: {
-                with: {
-                  profile: {
-                    with: {
-                      avatarImage: true,
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      });
+      const user = await getUserWithProfiles({ authUserId });
 
       if (!user) {
         throw new UnauthorizedError('User not found');
