+++ services/api/src/routers/organization/deletePost.ts
@@ -1,12 +1,9 @@
-import { UnauthorizedError, getOrgAccessUser } from '@op/common';
-import { posts, postsToOrganizations } from '@op/db/schema';
+import { UnauthorizedError, getOrgAccessUser, deletePostById } from '@op/common';
 import { TRPCError } from '@trpc/server';
-import { and, eq } from 'drizzle-orm';
 import type { OpenApiMeta } from 'trpc-to-openapi';
 import { z } from 'zod';
 
 import withAuthenticated from '../../middlewares/withAuthenticated';
-import withDB from '../../middlewares/withDB';
 import withRateLimited from '../../middlewares/withRateLimited';
 import { loggedProcedure, router } from '../../trpcFactory';
 
@@ -26,7 +23,6 @@ export const deletePost = router({
   deletePost: loggedProcedure
     .use(withRateLimited({ windowSize: 10, maxRequests: 5 }))
     .use(withAuthenticated)
-    .use(withDB)
     .meta(meta)
     .input(
       z.object({
@@ -39,7 +35,6 @@ export const deletePost = router({
     .output(z.object({ success: z.boolean() }))
     .mutation(async ({ input, ctx }) => {
       const { id, organizationId } = input;
-      const { db } = ctx.database;
 
       const user = await getOrgAccessUser({
         organizationId,
@@ -50,32 +45,16 @@ export const deletePost = router({
         throw new UnauthorizedError();
       }
 
-      // Verify the post exists and belongs to the organization
-      const postExists = await db
-        .select()
-        .from(posts)
-        .innerJoin(
-          postsToOrganizations,
-          eq(posts.id, postsToOrganizations.postId),
-        )
-        .where(
-          and(
-            eq(posts.id, id),
-            eq(postsToOrganizations.organizationId, organizationId),
-          ),
-        )
-        .limit(1);
-
-      if (!postExists.length) {
-        throw new TRPCError({
-          code: 'NOT_FOUND',
-          message:
-            'Post not found or does not belong to the specified organization',
-        });
+      try {
+        return await deletePostById({ postId: id, organizationId });
+      } catch (error) {
+        if (error instanceof Error && error.message.includes('Post not found')) {
+          throw new TRPCError({
+            code: 'NOT_FOUND',
+            message: error.message,
+          });
+        }
+        throw error;
       }
-
-      await db.delete(posts).where(eq(posts.id, id));
-
-      return { success: true };
     }),
 });
