+++ services/api/src/routers/organization/inviteUser.ts
@@ -1,12 +1,9 @@
-import { UnauthorizedError, sendInvitationEmail } from '@op/common';
-import { OPURLConfig } from '@op/core';
-import { allowList } from '@op/db/schema';
+import { UnauthorizedError, inviteUsersToOrganization } from '@op/common';
 import { TRPCError } from '@trpc/server';
 // import type { OpenApiMeta } from 'trpc-to-openapi';
 import { z } from 'zod';
 
 import withAuthenticated from '../../middlewares/withAuthenticated';
-import withDB from '../../middlewares/withDB';
 import withRateLimited from '../../middlewares/withRateLimited';
 import { loggedProcedure, router } from '../../trpcFactory';
 
@@ -60,14 +57,12 @@ export const inviteUserRouter = router({
     // Middlewares
     .use(withRateLimited({ windowSize: 60, maxRequests: 10 }))
     .use(withAuthenticated)
-    .use(withDB)
     // Router
     // .meta(meta)
     .input(inputSchema)
     .output(outputSchema)
     .mutation(async ({ ctx, input }) => {
       try {
-        const { db } = ctx.database;
         const { id: authUserId } = ctx.user;
 
         // Handle both single email and multiple emails input
@@ -77,129 +72,23 @@ export const inviteUserRouter = router({
         const targetOrganizationId = input.organizationId;
         const personalMessage = input.personalMessage;
 
-        // Get the current user's database record with organization details
-        const authUser = await db.query.users.findFirst({
-          where: (table, { eq }) => eq(table.authUserId, authUserId),
-          with: {
-            currentOrganization: {
-              with: {
-                profile: true,
-              },
-            },
-            currentProfile: true,
-          },
+        return await inviteUsersToOrganization({
+          emails: emailsToProcess,
+          role,
+          organizationId: targetOrganizationId,
+          personalMessage,
+          authUserId,
+          authUserEmail: ctx.user.email,
         });
-
-        // For new organization invites, we don't need the user to be in an organization
-        // For existing organization invites, we do need it
-        if (
-          (!authUser?.currentProfileId && !authUser?.lastOrgId) ||
-          (!authUser.currentOrganization && !authUser.currentProfile)
-        ) {
-          throw new UnauthorizedError(
-            'User must be associated with an organization to send invites',
-          );
-        }
-
-        const currentProfile =
-          authUser.currentProfile ??
-          (authUser.currentOrganization as any)?.profile;
-
-        const results = {
-          successful: [] as string[],
-          failed: [] as { email: string; reason: string }[],
-        };
-
-        // Process each email
-        for (const rawEmail of emailsToProcess) {
-          const email = rawEmail.toLowerCase();
-          try {
-            // Check if email is already in the allowList
-            const existingEntry = await db.query.allowList.findFirst({
-              where: (table, { eq }) => eq(table.email, email),
-            });
-
-            if (!existingEntry) {
-              // Determine metadata based on whether it's a new organization invite
-              const metadata = targetOrganizationId
-                ? {
-                    invitedBy: authUserId,
-                    invitedAt: new Date().toISOString(),
-                    inviteType: 'new_organization',
-                    personalMessage: personalMessage,
-                    inviterOrganizationName:
-                      (currentProfile as any)?.profile?.name || 'Common',
-                  }
-                : {
-                    invitedBy: authUserId,
-                    invitedAt: new Date().toISOString(),
-                    personalMessage: personalMessage,
-                    role,
-                  };
-
-              // Add the email to the allowList
-              await db.insert(allowList).values({
-                email,
-                organizationId: targetOrganizationId ?? null,
-                metadata,
-              });
-            }
-
-            // Send invitation email
-            try {
-              await sendInvitationEmail({
-                to: email,
-                inviterName:
-                  authUser?.name || ctx.user.email || 'A team member',
-                organizationName: targetOrganizationId
-                  ? (authUser?.currentOrganization as any)?.profile?.name ||
-                    'an organization'
-                  : undefined,
-                inviteUrl: OPURLConfig('APP').ENV_URL,
-                message: personalMessage,
-              });
-              results.successful.push(email);
-            } catch (emailError) {
-              console.error(
-                `Failed to send invitation email to ${email}:`,
-                emailError,
-              );
-              // Email failed but database insertion succeeded
-              results.successful.push(email);
-            }
-          } catch (error) {
-            console.error(`Failed to process invitation for ${email}:`, error);
-            results.failed.push({
-              email,
-              reason: error instanceof Error ? error.message : 'Unknown error',
-            });
-          }
-        }
-
-        const totalEmails = emailsToProcess.length;
-        const successCount = results.successful.length;
-
-        let message: string;
-        if (successCount === totalEmails) {
-          message = `All ${totalEmails} invitation${totalEmails > 1 ? 's' : ''} sent successfully`;
-        } else if (successCount > 0) {
-          message = `${successCount} of ${totalEmails} invitations sent successfully`;
-        } else {
-          message = 'No invitations were sent successfully';
-        }
-
-        return {
-          success: successCount > 0,
-          message,
-          details: {
-            successful: results.successful,
-            failed: results.failed,
-          },
-        };
       } catch (error) {
         // Re-throw TRPCError as-is
         if (error instanceof TRPCError) {
           throw error;
+        }
+
+        // Handle specific errors
+        if (error instanceof Error && error.message.includes('User must be associated')) {
+          throw new UnauthorizedError(error.message);
         }
 
         // Handle other errors
