+++ services/api/src/routers/organization/uploadAvatarImage.ts
@@ -1,13 +1,12 @@
+import { trackImageUpload } from '@op/analytics';
 import { createServerClient } from '@op/supabase/lib';
 import { TRPCError } from '@trpc/server';
 import { waitUntil } from '@vercel/functions';
 import { Buffer } from 'buffer';
 import type { OpenApiMeta } from 'trpc-to-openapi';
 import { z } from 'zod';
-import { trackImageUpload } from '@op/analytics';
 
 import withAuthenticated from '../../middlewares/withAuthenticated';
-import withDB from '../../middlewares/withDB';
 import withRateLimited from '../../middlewares/withRateLimited';
 import { loggedProcedure, router } from '../../trpcFactory';
 import { MAX_FILE_SIZE, sanitizeS3Filename } from '../../utils';
@@ -37,7 +36,6 @@ export const uploadAvatarImage = router({
     // middlewares
     .use(withRateLimited({ windowSize: 10, maxRequests: 10 }))
     .use(withAuthenticated)
-    .use(withDB)
     // Router
     .meta(meta)
     .input(
@@ -149,11 +147,11 @@ export const uploadAvatarImage = router({
       logger.info(
         'RETURNING UPLOAD URL' + signedUrlData.signedUrl + ' - ' + filePath,
       );
-      
+
       // Track analytics - for organization uploads, we'll track as new uploads since they're temporary (non-blocking)
       const imageType = filePath.includes('banner') ? 'banner' : 'profile';
       waitUntil(trackImageUpload(ctx.user.id, imageType, false));
-      
+
       return {
         url: signedUrlData.signedUrl,
         path: filePath,
